requirements.txt
requests==2.32.3
pandas==2.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dbt-postgres==1.11.2
